requirements.txt
torch
pandas
numpy
scikit-learn
pytest
pyyaml